feat(ObjectDetectionModel.py): add ObjectDetectionModel class and non_max_suppression function feat(__init__.py): import ObjectDetectionModel class The ObjectDetectionModel class is added to the min_max_models package. It contains a YOLOv7 model and a non_max_suppression function to perform object detection on an image. The non_max_suppression function filters out overlapping bounding boxes and returns the count of non-overlapping boxes. The __init__.py file is updated to import the ObjectDetectionModel class.
feat(yolov7_utils.py): add utility functions for loading and tracing YOLOv7 models
This commit adds several utility functions for loading and tracing YOLOv7 models. The `TracedModel` class is used to convert a YOLOv7 model to a traced model, which can be used for faster inference. The `attempt_download` function is used to download the YOLOv7 model weights if they are not already present. The `Ensemble` class is used to create an ensemble of YOLOv7 models. The `Conv` class is used to define a standard convolution layer. The `autopad` function is used to pad the input to a convolution layer. The `attempt_load` function is used to load an ensemble of YOLOv7 models from a set of weights.